Settings UI injection logic move to `OptionsPanel`
